resolved requested changes in PR
